[refactoring] 전체적으로 구조 개선...
