+++ .eslintrc.cjs
@@ -1,15 +1,31 @@
 module.exports = {
-	root: true,
-	extends: ['eslint:recommended', 'prettier'],
-	plugins: ['svelte3'],
-	overrides: [{ files: ['*.svelte'], processor: 'svelte3/svelte3' }],
-	parserOptions: {
-		sourceType: 'module',
-		ecmaVersion: 2020
+	'env': {
+		'browser': true,
+		'es2021': true
 	},
-	env: {
-		browser: true,
-		es2017: true,
-		node: true
+	'extends': 'eslint:recommended',
+	'overrides': [
+	],
+	'parserOptions': {
+		'ecmaVersion': 'latest',
+		'sourceType': 'module'
+	},
+	'rules': {
+		'indent': [
+			'error',
+			'tab'
+		],
+		'linebreak-style': [
+			'error',
+			'unix'
+		],
+		'quotes': [
+			'error',
+			'single'
+		],
+		'semi': [
+			'error',
+			'always'
+		]
 	}
 };
