--- src/routes/scripts/localStorage.js
@@ -1,15 +1,0 @@
-import { showThisLocation } from "./store";
-
-export function getLocation() {
-	const loc = window.localStorage.getItem('location')
-	if (loc !== undefined) {
-		return loc;
-	} else {
-		return 'gaepo';
-	}
-}
-
-export function setLocation(loc) {
-	showThisLocation.set(loc);
-	window.localStorage.setItem('location', loc);
-}
