--- src/routes/scripts/store.js
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store';
-
-export const restaurantDatas = writable([]);
-export const showThisLocation = writable();
